UCP/FLUSH: Address selective flush review comments
