+++ test/gtest/ucp/test_ucp_fence.cc
@@ -6,107 +6,6 @@
 
 #include "ucp_test.h"
 
-extern "C" {
-#include <ucp/core/ucp_request.inl>
-#include <ucp/rma/rma.h>
-#include <uct/base/uct_iface.h>
-}
-
-class test_ucp_flush_lane_state : public ucs::test {
-};
-
-UCS_TEST_F(test_ucp_flush_lane_state, replace_unstarted_lane)
-{
-    ucp_lane_map_t all_lanes     = UCS_BIT(0) | UCS_BIT(1);
-    ucp_lane_map_t lane_mask     = UCS_BIT(1);
-    ucp_lane_map_t started_lanes = 0;
-    int count_diff;
-
-    count_diff = ucp_ep_flush_lane_state_update(
-            UCS_BIT(0) | UCS_BIT(2), 0, &started_lanes, &all_lanes,
-            &lane_mask);
-
-    EXPECT_EQ(0, count_diff);
-    EXPECT_EQ(UCS_BIT(0) | UCS_BIT(2), all_lanes);
-    EXPECT_EQ(UCS_BIT(1) | UCS_BIT(2), lane_mask);
-}
-
-UCS_TEST_F(test_ucp_flush_lane_state, replace_started_lane)
-{
-    ucp_lane_map_t all_lanes     = UCS_BIT(0) | UCS_BIT(1);
-    ucp_lane_map_t lane_mask     = UCS_BIT(1);
-    ucp_lane_map_t started_lanes = UCS_BIT(1);
-    int count_diff;
-
-    count_diff = ucp_ep_flush_lane_state_update(
-            UCS_BIT(0) | UCS_BIT(2), 0, &started_lanes, &all_lanes,
-            &lane_mask);
-
-    EXPECT_EQ(1, count_diff);
-    EXPECT_EQ(UCS_BIT(0) | UCS_BIT(2), all_lanes);
-    EXPECT_EQ(UCS_BIT(1) | UCS_BIT(2), lane_mask);
-}
-
-UCS_TEST_F(test_ucp_flush_lane_state, replace_same_index_lane)
-{
-    ucp_lane_map_t all_lanes     = UCS_BIT(0) | UCS_BIT(1);
-    ucp_lane_map_t lane_mask     = UCS_BIT(0);
-    ucp_lane_map_t started_lanes = UCS_BIT(0);
-    int count_diff;
-
-    count_diff = ucp_ep_flush_lane_state_update(
-            UCS_BIT(0) | UCS_BIT(1), 1, &started_lanes, &all_lanes,
-            &lane_mask);
-
-    EXPECT_EQ(1, count_diff);
-    EXPECT_EQ(0, started_lanes);
-    EXPECT_EQ(UCS_BIT(0) | UCS_BIT(1), all_lanes);
-    EXPECT_EQ(UCS_BIT(0) | UCS_BIT(1), lane_mask);
-}
-
-UCS_TEST_F(test_ucp_flush_lane_state, destroyed_started_lane_is_not_unstarted)
-{
-    EXPECT_FALSE(ucp_ep_flush_has_unstarted_lanes(
-            UCS_BIT(0) | UCS_BIT(2),
-            UCS_BIT(0) | UCS_BIT(1) | UCS_BIT(2)));
-}
-
-UCS_TEST_F(test_ucp_flush_lane_state, live_unstarted_lane)
-{
-    EXPECT_TRUE(ucp_ep_flush_has_unstarted_lanes(
-            UCS_BIT(0) | UCS_BIT(2), UCS_BIT(0) | UCS_BIT(1)));
-}
-
-static unsigned test_flush_call_count;
-
-static ucs_status_t
-test_flush_count_calls(uct_ep_h, unsigned, uct_completion_t *)
-{
-    ++test_flush_call_count;
-    return UCS_ERR_IO_ERROR;
-}
-
-static uct_completion_t *test_flush_comp;
-static uct_ep_h test_flush_eps[2];
-
-static ucs_status_t
-test_flush_inprogress(uct_ep_h ep, unsigned, uct_completion_t *comp)
-{
-    if (test_flush_call_count < ucs_static_array_size(test_flush_eps)) {
-        test_flush_eps[test_flush_call_count] = ep;
-    }
-
-    ++test_flush_call_count;
-    test_flush_comp = comp;
-    return UCS_INPROGRESS;
-}
-
-static void test_flush_completion(ucp_request_t *req)
-{
-    ucp_request_complete_send(req, req->status);
-}
-
-
 class test_ucp_fence : public ucp_test {
 public:
     virtual void init() {
@@ -278,209 +177,6 @@ public:
 UCS_TEST_P(test_ucp_fence32, atomic_add_fadd) {
     test<uint32_t>(&test_ucp_fence32::blocking_add<uint32_t>,
                    &test_ucp_fence32::blocking_fadd<uint32_t>);
-}
-
-UCS_TEST_P(test_ucp_fence32, empty_lane_mask_skips_transport_flush)
-{
-    ucp_request_param_t param = {};
-    ucp_ep_h ep;
-    uct_iface_h iface;
-    uct_ep_flush_func_t flush_func;
-    ucs_status_ptr_t request;
-
-    if (!is_self()) {
-        UCS_TEST_SKIP_R("Direct flush interception requires self transport");
-    }
-
-    sender().connect(&receiver(), get_ep_params());
-    ep         = sender().ep();
-    iface      = ucp_ep_get_lane(ep, 0)->iface;
-    flush_func = iface->ops.ep_flush;
-
-    test_flush_call_count = 0;
-    iface->ops.ep_flush   = test_flush_count_calls;
-
-    UCP_WORKER_THREAD_CS_ENTER_CONDITIONAL(ep->worker);
-    request = ucp_ep_flush_lanes_internal(
-            ep, 0, &param, NULL, test_flush_completion, "flush_lane_mask_test",
-            UCT_FLUSH_FLAG_LOCAL, 0);
-    UCP_WORKER_THREAD_CS_EXIT_CONDITIONAL(ep->worker);
-
-    iface->ops.ep_flush = flush_func;
-
-    EXPECT_EQ(0, test_flush_call_count);
-    EXPECT_EQ(NULL, request);
-
-    disconnect(sender());
-    disconnect(receiver());
-}
-
-UCS_TEST_P(test_ucp_fence32, replace_lane_during_selective_flush)
-{
-    ucp_request_param_t param = {};
-    uct_ep_t replacement_lane = {};
-    ucp_ep_h ep;
-    uct_ep_h original_lane;
-    uct_iface_h iface;
-    uct_ep_flush_func_t flush_func;
-    ucs_status_ptr_t request;
-
-    if (!is_self()) {
-        UCS_TEST_SKIP_R("Direct flush interception requires self transport");
-    }
-
-    sender().connect(&receiver(), get_ep_params());
-    ep                       = sender().ep();
-    original_lane            = ucp_ep_get_lane(ep, 0);
-    iface                    = original_lane->iface;
-    replacement_lane.iface   = iface;
-    flush_func               = iface->ops.ep_flush;
-    test_flush_call_count    = 0;
-    test_flush_comp          = NULL;
-    test_flush_eps[0]        = NULL;
-    test_flush_eps[1]        = NULL;
-    iface->ops.ep_flush      = test_flush_inprogress;
-
-    UCP_WORKER_THREAD_CS_ENTER_CONDITIONAL(ep->worker);
-    request = ucp_ep_flush_lanes_internal(
-            ep, 0, &param, NULL, test_flush_completion,
-            "replace_lane_during_flush", UCT_FLUSH_FLAG_LOCAL, UCS_BIT(0));
-    EXPECT_TRUE(UCS_PTR_IS_PTR(request));
-    EXPECT_EQ(1u, test_flush_call_count);
-    EXPECT_EQ(original_lane, test_flush_eps[0]);
-
-    ucp_ep_set_lane(ep, 0, &replacement_lane);
-    EXPECT_TRUE(test_flush_comp != NULL);
-    if (test_flush_comp != NULL) {
-        uct_invoke_completion(test_flush_comp, UCS_OK);
-    }
-
-    EXPECT_EQ(2u, test_flush_call_count);
-    EXPECT_EQ(&replacement_lane, test_flush_eps[1]);
-
-    EXPECT_TRUE(test_flush_comp != NULL);
-    if (test_flush_comp != NULL) {
-        uct_invoke_completion(test_flush_comp, UCS_OK);
-    }
-    ucp_ep_set_lane(ep, 0, original_lane);
-    iface->ops.ep_flush = flush_func;
-    UCP_WORKER_THREAD_CS_EXIT_CONDITIONAL(ep->worker);
-
-    EXPECT_UCS_OK(ucp_request_check_status(request));
-    ucp_request_release(request);
-
-    disconnect(sender());
-    disconnect(receiver());
-}
-
-UCS_TEST_P(test_ucp_fence32, lane_generation_updates_on_pointer_change)
-{
-    uint64_t lane_generation;
-    ucp_ep_h ep;
-    uct_ep_h lane;
-
-    if (!is_self()) {
-        UCS_TEST_SKIP_R("Direct lane-state test requires self transport");
-    }
-
-    sender().connect(&receiver(), get_ep_params());
-    ep              = sender().ep();
-    lane_generation = ep->ext->lane_generation;
-    lane            = ucp_ep_get_lane(ep, 0);
-
-    ucp_ep_set_lane(ep, 0, lane);
-    EXPECT_EQ(lane_generation, ep->ext->lane_generation);
-
-    ucp_ep_set_lane(ep, 0, NULL);
-    EXPECT_EQ(lane_generation + 1, ep->ext->lane_generation);
-
-    ucp_ep_set_lane(ep, 0, lane);
-    EXPECT_EQ(lane_generation + 2, ep->ext->lane_generation);
-
-    disconnect(sender());
-    disconnect(receiver());
-}
-
-UCS_TEST_P(test_ucp_fence32, recycled_ep_lane_storage_initialization)
-{
-    ucp_worker_h worker = sender().worker();
-    ucp_ep_h ep;
-    ucp_ep_h recycled_ep;
-    ucp_lane_index_t lane;
-
-    UCS_ASYNC_BLOCK(&worker->async);
-    ASSERT_UCS_OK(ucp_ep_create_base(worker, UCP_EP_INIT_FLAG_INTERNAL,
-                                     "lane-init", "lane-init", &ep));
-
-    for (lane = 0; lane < UCP_MAX_FAST_PATH_LANES; ++lane) {
-        ep->uct_eps[lane] = reinterpret_cast<uct_ep_h>(
-                static_cast<uintptr_t>(lane + 1));
-    }
-
-    ucp_ep_delete(ep);
-
-    ASSERT_UCS_OK(ucp_ep_create_base(worker, UCP_EP_INIT_FLAG_INTERNAL,
-                                     "lane-recycle", "lane-recycle",
-                                     &recycled_ep));
-
-    EXPECT_EQ(ep, recycled_ep);
-    EXPECT_EQ(0, recycled_ep->ext->lane_generation);
-    for (lane = 0; lane < UCP_MAX_FAST_PATH_LANES; ++lane) {
-        EXPECT_EQ(NULL, recycled_ep->uct_eps[lane]);
-    }
-
-    ucp_ep_delete(recycled_ep);
-    UCS_ASYNC_UNBLOCK(&worker->async);
-}
-
-UCS_TEST_P(test_ucp_fence32, slow_lane_storage_initialization)
-{
-    const unsigned num_lanes = UCP_MAX_FAST_PATH_LANES + 2;
-    ucp_worker_h worker       = sender().worker();
-    ucp_ep_h ep;
-    ucp_lane_index_t lane;
-
-    UCS_ASYNC_BLOCK(&worker->async);
-    ASSERT_UCS_OK(ucp_ep_create_base(worker, UCP_EP_INIT_FLAG_INTERNAL,
-                                     "lane-init", "lane-init", &ep));
-    ASSERT_UCS_OK(ucp_ep_realloc_lanes(ep, num_lanes));
-
-    for (lane = UCP_MAX_FAST_PATH_LANES; lane < num_lanes; ++lane) {
-        ep->ext->uct_eps[lane - UCP_MAX_FAST_PATH_LANES] =
-                reinterpret_cast<uct_ep_h>(static_cast<uintptr_t>(lane + 1));
-    }
-
-    ep->ext->lane_generation = 0;
-    ASSERT_UCS_OK(ucp_ep_realloc_lanes(ep, num_lanes));
-
-    EXPECT_EQ(0, ep->ext->lane_generation);
-    for (lane = 0; lane < num_lanes; ++lane) {
-        EXPECT_EQ(NULL, ucp_ep_get_lane(ep, lane));
-    }
-
-    ucp_ep_delete(ep);
-    UCS_ASYNC_UNBLOCK(&worker->async);
-}
-
-UCS_TEST_P(test_ucp_fence32, fast_lane_storage_initialization_tracks_change)
-{
-    const unsigned num_lanes = UCP_MAX_FAST_PATH_LANES + 1;
-    ucp_worker_h worker       = sender().worker();
-    ucp_ep_h ep;
-
-    UCS_ASYNC_BLOCK(&worker->async);
-    ASSERT_UCS_OK(ucp_ep_create_base(worker, UCP_EP_INIT_FLAG_INTERNAL,
-                                     "lane-init", "lane-init", &ep));
-
-    ep->uct_eps[0]           = reinterpret_cast<uct_ep_h>(1);
-    ep->ext->lane_generation = 0;
-    ASSERT_UCS_OK(ucp_ep_realloc_lanes(ep, num_lanes));
-
-    EXPECT_EQ(1, ep->ext->lane_generation);
-    EXPECT_EQ(NULL, ep->uct_eps[0]);
-
-    ucp_ep_delete(ep);
-    UCS_ASYNC_UNBLOCK(&worker->async);
 }
 
 UCP_INSTANTIATE_TEST_CASE(test_ucp_fence32)
